+++ libs/ECS/Components/LifeComponent.hpp
@@ -16,11 +16,11 @@ namespace ecs
     class Life : public Component{
       public:
         /// @brief The life's point of the entity at the beginning
-        unsigned int lifePoint;
+        unsigned short lifePoint;
 
         /// @brief Constructor of the Life Component
         /// @param newLifePoint the life given to the entity
-        Life(unsigned int newLifePoint) : lifePoint(newLifePoint) {};
+        Life(unsigned short newLifePoint) : lifePoint(newLifePoint) {};
         /// @brief Copy Constructor of the Life component
         /// @param old The copy of the Life component
         Life(const Life &old) : lifePoint(old.lifePoint) {};
